+++ tests/examples/test_scripts.py
@@ -95,6 +95,11 @@ def test_16():  # noqa: D103
     clean_example(['CH4_fods.xyz'])
 
 
+@pytest.mark.slow()
+def test_17():  # noqa: D103
+    execute_example('17_reduced_density_gradient')
+
+
 if __name__ == '__main__':
     file_path = pathlib.Path(inspect.stack()[0][1])
     pytest.main(file_path)